Monky

Description

This module contains the main logic for monky. This has to be included in the Monky.hs. The entry point for monky is startLoop which has to be called with a [IO Modules]. This type was chosen to make the syntax in the config file prettier.

The submodules of this generally provide an interface based on handles. To use them, get the handle at the beginning of you application and hand it to other functions later on.

Documentation

startLoop :: [IO Modules] -> IO ()

Starts the main loop for monky

This is the entry-point for printing normal lines. This function never returns.

getVersion :: (Int, Int, Int, Int)

Export the version of monky to modules