Monky.Blkid

Description

This module allows access to libblkid functionality

Since the library does not have to exist on every system we link against it on runtime when needed. We don't have a handle or something cool like that, so we will just load it when needed und unload again after

Documentation

evaluateTag :: String -> String -> IO (Maybe String)

Evaluate a tag

evaluateSpec :: String -> IO (Maybe String)

Evaluate a spec

evaluateTag' :: String -> String -> LibBlkid -> IO (Maybe String)

Version to reuse the LibBlkid

evaluateSpec' :: String -> LibBlkid -> IO (Maybe String)

version to reuse the LibBlkid

withLibBlkid :: (LibBlkid -> IO a) -> IO a

Execute an IO action with an instance of LibBlkid