Monky.CPU

Description

Documentation

data CPUHandle

The handle exported by this module

newtype Numa

Handle used for by numa aware version

Constructors

Numa [NumaHandle]

data NumaHandle

Numa aware version of CPUHandle

Constructors

NumaHandle

Fields

numaCpus :: [String]

numaHandle :: CPUHandle

getCPUHandle :: ScalingType -> IO CPUHandle

Version for getCPUHandle' that defaults to thermal zone thermal_zone0

getCPUHandle'

Arguments

:: ScalingType
The scaling type, either ScalingMax or ScalingCur

-> Maybe String
The thermal zone to use or Nothing if there isn't any

-> IO CPUHandle

Create an CPUHandle

getCPUPercent :: CPUHandle -> IO [Int]

Get the cpu usage in percent for each (virtual) cpu

getNumaPercent :: NumaHandle -> IO [Int]

Read node information

getCPUTemp :: CPUHandle -> IO Int

get current CPU temperature

getCPUMaxScalingFreq :: CPUHandle -> IO Float

This function returns a frequency according the ScalingType of the handle.

The returned valued will be the max of all (virtual) proceessors on the system.

data ScalingType

Which values should be returned by getCPUFreq

Constructors

ScalingMax
Use the maximum frequencie allowed

ScalingCur
Use the current frequencie used

ScalingNone
Don't open any files for scaling type

getNumaHandles :: ScalingType -> IO Numa

getnumaHandles' but try to guess the thermal zones