Monky.Connectivity

Description

This module checks periodically if the current system can establish a network connection (TCP) to a given host on a given port. It does not care about reject, it is intended to test whether a firewall drops packages or a (tethered) connection is stable.

Documentation

data ConnHandle

The handle exposed by this module

getConnH

Arguments

:: String
The Host to use for connectivity probing

-> Int
Which port to use for connecivity probing (drop is bad)

-> IO ConnHandle

Get a handle to check for connectivity

hasConn :: ConnHandle -> IO Bool

Get the current connected state from the handle