Monky.Disk.Btrfs

Description

This module allows for some support for btrfs devices. This may be renamed in the future when a general block-device module appears.

Documentation

data BtrfsHandle

The FsInfo exported by this module

getBtrfsHandle

Arguments

:: String
The UUID of the file system to monitor

-> IO (Maybe (BtrfsHandle, [String]))

Try to create a btfshanlde given the UUID

This will create a BtrfsHandle which is an instance of FsInfo and a list of block devices that are slaves of our file system.

This allows the upper layer to monitor the read/write rates of all block devices that belong to our file system and report them as read/write rate for the file system.

Due to compression and encryption the read/write rate on the block device may be quite different to the one that application see.