Monky.Examples.CPU

Description

Documentation

getCPUHandle

Arguments

:: ScalingType
The type of scaling frequency that should be reported

-> String
The thermal zone of the cpu

-> IO CPUHandle

Get a CPUHandle

This is a shiny combination of the raw handles exported by this module

getCPUHandle' :: ScalingType -> IO CPUHandle

Same as getCPUHandle but tries to guess the thermal zone

getNumaHandles

Arguments

:: ScalingType
The type of scaling frequency that should be reported

-> [String]
A list of thermal zones for our numa handles

-> IO NumaHandle

Numa aware version of getCPUHandle

getNumaHandles' :: ScalingType -> IO NumaHandle

Same as getNumaHandles but tries to guess the thermal zone

data ScalingType

Which values should be returned by getCPUFreq

Constructors

ScalingMax
Use the maximum frequencie allowed

ScalingCur
Use the current frequencie used

ScalingNone
Don't open any files for scaling type

data CPUHandle

The handle type for the default setup

data NumaHandle

The handle type for the default setup (numa aware)

data FreqHandle

Handle to access cpu frequency information

data TempHandle

Handle to acces thermal zone information

getFreqHandle :: ScalingType -> IO FreqHandle

Get a handle to access max node frequency

getFreqNuma :: ScalingType -> RawNuma -> IO FreqHandle

Numa away version of getFreqHandle

getTempHandle :: String -> IO TempHandle

Get a TempHandle for the specified thermal zone

getTempHandle' :: IO TempHandle

Get the TempHandle for the system (tries to guess CPU zone)

getTempHandles :: IO [TempHandle]

Get a TempHandle for each cpu termal zone on the system. will be infinite list, with invalidated zones for convinience

getRawNumas :: IO [RawNuma]

Get raw numa aware cpu handles

getRawCPU :: IO RawCPU

Get a raw CPU handle