Monky.Examples.Memory

Description

For older kernels the memory available value may be wrong. When not provided by the kernel it's computed as `free + cached + buffers`. If you find this to be to inacurate, use the version for memory free. But be aware, that free memory on linux can go close to zero because of io buffers.

Documentation

getMemoryHandle :: IO MHandle

Get the the memory handle (available)

getMemoryBarHandle

Arguments

:: Float
A factor to modify bar length. Total-length: 100/f

-> IO MBHandle

Get the MBHandle

getMemoryFreeHandle :: IO MFHandle

Get the the memory handle (free)

data MHandle

Simple handle to display current memory available

data MFHandle

Simple handle to display current free memory

data MBHandle

Handle to display the current memory usage (usedcachesfree) as horizontal bar