Monky.Examples.Modify

Documentation

data ModifyHandle a

Handle for duing pure modification of module output

getModifyHandle :: ([MonkyOut] -> [MonkyOut]) -> IO a -> IO (ModifyHandle a)

Get a handle to purely modify another modules output

data IOModifyHandle a

Handle for modifying output in IO monad

getIOModifyHandle :: ([MonkyOut] -> IO [MonkyOut]) -> IO a -> IO (IOModifyHandle a)

Get a handle that can modify another handles output in the IO monad