Monky.Examples.Network

Description

Documentation

getNetworkHandles'

Arguments

:: Text
The string to use when the network is disconnected

-> [String]
The Network adapters to monitor

-> IO NetworkHandles'

Actually get the network handle

getStaticHandle

Arguments

:: Text
The string to use when network is disconnected

-> String
Name of the network interface

-> IO NetworkStaticHandle

Get a handle for a static network interface

getDynamicHandle

Arguments

:: Text
The string to use when network is disconnected

-> (String -> Bool)
The filter function for the dynamic handle

-> IO NetworkDynHandle

Get a handle that will update with new or disappearing network interfaces

data NetworkDynHandle

Handle for dynamic network handling

data NetworkStaticHandle

HAndle for a staticly configured network interface