Monky.Examples.Prepend

Description

The *PrependAppend functions can be used instead of pollPackevtPack. If you want to chain a prepended module into something else, use the raw Constructors. packPrepend [MonkyPlain CPU] 1 $ getCPUHandle ScalingCur pack 1 $ Prep [MonkyPlain CPU] $ getCPUHandle ScalingCur

Documentation

data PrepHandle

The handle used by this module, contains underlying module and string

Constructors

forall m . PollModule m => Prep [MonkyOut] m

data PostHandle

The handle used by this module, contains underlying module and string

Constructors

forall m . PollModule m => Post [MonkyOut] m

packPrepend

Arguments

:: PollModule a

=> [MonkyOut]
The String to prepend

-> Int
The refresh rate for this module

-> IO a
The function to get the module

-> IO Modules
The returned handle

Create a module that should be prepended with some string

This allows you to prepend an instance of a module with a fixed String.

For usage look at pollPack.

packAppend

Arguments

:: PollModule a

=> [MonkyOut]
The String to prepend

-> Int
The refresh rate for this module

-> IO a
The function to get the module

-> IO Modules
The returned handle

Create a module that should be appended with some string

This allows you to append an instance of a module with a fixed String.

For usage look at pollPack.

data EvtPrepHandle

EvtModule prepend type

Constructors

forall m . EvtModule m => EvtPrep [MonkyOut] m

data EvtPostHandle

EvtModule append type

Constructors

forall m . EvtModule m => EvtPost [MonkyOut] m

evtPrepend :: EvtModule a => [MonkyOut] -> IO a -> IO Modules

packPrepend for EvtModules

evtAppend :: EvtModule a => [MonkyOut] -> IO a -> IO Modules

packAppend for EvtModules