Monky.Examples.Wifi

Description

Documentation

getWifiHandle

Arguments

:: [WifiFormat]
Format String for output generation

-> Text
Text that should be displayed when wifi is disconnected

-> String
Name of the interface

-> IO WifiHandle

Get a wifi handle

data WifiHandle

The handle type for this module

data WifiFormat

A typesafe version of a format string

Constructors

FormatChannel
Print the current networks channel

FormatRates
Print the current network max supported rate (always 54Mbit/s for me)

FormatName
Print the ESSID of the current network, may look weird because SSIDs are

FormatFreq
Print the frequency the current network sends on (related to channel)

FormatMBM
I think something something signal strength

FormatText Text
Print a plaintext string