Monky.IP.Raw

Description

This may change at any time when the main IP module changes. Consider this API unstable!

Documentation

data IP

Datatype for IP addresses, abstracts over v4/v6

Constructors

IPv4 IP4

IPv6 IP6

data IP4

IPv4 addresses

data IP6

IPv6 addresses

parseIP :: String -> IP4

Parse an IP4 from a String

ipFromBS :: ByteString -> IP

Read an IP from a ByteString. The type is determined by the size of the ByteString.

familyToNum :: Num a => AddressFamily -> a

Get the number associated with the family address. This is for interfacing with libraries

data AddressFamily

AddressFamilies support for libraries

Constructors

AF_UNSPEC

AF_INET

AF_INET6

getAddrFamily :: IP -> AddressFamily

Get the address family for a given ip address