Monky.MPD

Description

This module creates a persistant connection to an mpd server and allows to query information or react to events waited for with MPDs idle

Documentation

data MPDSocket

A type safe socket for this module, basically a handle

data State

The current state of the MPD player submodule

Constructors

Playing
Playing

Stopped
Stopped

Paused
Paused

data TagCollection

Collection of tags MPD supports

Constructors

TagCollection

Fields

tagArtist :: Maybe String

tagArtistSort :: Maybe String

tagAlbum :: Maybe String

tagAlbumSort :: Maybe String

tagAlbumArtist :: Maybe String

tagAlbumArtistSort :: Maybe String

tagTitle :: Maybe String

tagTrack :: Maybe String

tagName :: Maybe String

tagGenre :: Maybe String

tagDate :: Maybe String

tagComposer :: Maybe String

tagPerformer :: Maybe String

tagComment :: Maybe String

tagDisc :: Maybe String

tagMArtistid :: Maybe String

tagMAlbumid :: Maybe String

tagMAlbumArtistid :: Maybe String

tagMTrackid :: Maybe String

tagMReleaseTrackid :: Maybe String

data SongInfo

Information about an song

Constructors

SongInfo

Fields

songFile :: String

songRange :: Maybe (Float, Float)

songMTime :: Maybe String

songTime :: Maybe Int

songDuration :: Maybe Float

songTags :: TagCollection

songPos :: Maybe Int

songInfoId :: Maybe Int

songPriority :: Maybe Int

data Status

Haskell type for the data returned by MPDs status command

Constructors

Status

Fields

volume :: Int

repeats :: Bool

random :: Bool

single :: Maybe Bool

consume :: Maybe Bool

playlist :: Int

playlistLength :: Int

state :: State

mixrAmpdb :: Float

xfade :: Maybe Int

mixrAmpDelay :: Maybe Int

song :: Maybe Int

songId :: Maybe Int

nextSong :: Maybe Int

nextSongId :: Maybe Int

time :: Maybe Int

elapsed :: Maybe Float

bitrate :: Maybe Int

duration :: Maybe Int

audio :: Maybe (Int, Int, Int)

updating :: Maybe Int

mpderror :: Maybe String

getMPDStatus :: MPDSocket -> IO (Either String Status)

Get the current status or an error from the MPD server

getMPDSong :: MPDSocket -> IO (Either String SongInfo)

Get the information about the song currently beeing player from the MPD server

getMPDSocket

Arguments

:: String
The host the server is on

-> String
The port the server listens on

-> IO (Either String MPDSocket)

Get a new MPDSocket

closeMPDSocket :: MPDSocket -> IO ()

Close the MPDSocket when it is no longer required

getMPDFd :: MPDSocket -> IO Fd

Get the raw Fd from the MPDSocket for eventing api

readOk :: MPDSocket -> IO (Either String ())

Get the OK sent by mpd after idle out of the pipe

goIdle

Arguments

:: MPDSocket

-> String
The string that should be send with idle (list of subsystems)

-> IO (Either String ())

Go into MPDs idle mode, this does return, but MPD wont time us out and will notify us if something happens

doQuery :: MPDSocket -> String -> ExceptT String IO [String]

Send a query to the MPD server and return a list of answers (line by line)

This does not filter out errors, error checking has to be done by the user. This does filter out the last OK though