Monky.Memory

Description

Documentation

data MemoryHandle

The memory handle used for all functions

getMemoryAvailable :: MemoryHandle -> IO Int

Return the memory available to processes

getMemoryHandle :: IO MemoryHandle

Get a memory handle

getMemoryTotal :: MemoryHandle -> IO Int

Get the total amount of memory in the system

getMemoryUsed :: MemoryHandle -> IO Int

Get the amount of memory used by the kernel and processes