Monky.Modules

Description

This module provides the Module class which is used to define a Monky compatible module.

Documentation

data Modules

A wrapper around module instances so they can be put into a list.

Constructors

forall a . Module a => MW a Int

class Module a where

The type class for modules

Methods

getText

Arguments

:: String
The current user

-> a
The handle to this module

-> IO String
The text segment that should be displayed for this module

getFDs

Arguments

:: a
The handle to this module

-> IO [Fd]
The Fds to listen on for events

getEventText

Arguments

:: Fd
The fd that triggered an event

-> String
The current user

-> a
The handle to this module

-> IO String
The text segment that should be displayed

This function is used instead of getText for event triggerd updates.

The default implementation mappes this to getText

setupModule :: a -> IO Bool

This is supposed to set up the module This is only needed, if the module may fail so doing nothing should be ok for all *normal* modules

The module should return True if the setup was successful (is usable) or False if it failed to enter the broken state right at startup

Do to the current inner working of monky, recoverModule will be called before setupModule can be called, which will make this unnecessary for most modules

getTextFailable :: String -> a -> IO (Maybe String)

This function is a wrapper around getText that allows modules to report a fail and go into a failed state

getEventTextFailable :: Fd -> String -> a -> IO (Maybe String)

This function is a wrapper around getEventText that allows modules to report a fail and go into a failed state

recoverModule :: a -> IO Bool

If a module failed earlier this will be called in periodicly until the module returns true to indicate a successful recovery

pack

Arguments

:: Module a

=> Int
The refresh rate for this module

-> IO a
The function to get a module (get??Handle)

-> IO Modules
The packed module ready to be given to startLoop

Function to make packaging modules easier