Monky.Network

Description

This module allows to add multiple network interfaces. If multiple network interfaces are active the first one in the list will be used.

Documentation

data NetworkHandles

The handle exported by this module

getReadWriteMulti :: NetworkHandles -> IO (Maybe (Int, Int))

Get the read/write rate of the first interface in the list that is connected

The value will be a tuple (Read, Write) in bit/s or Nothing if no network is connected.

getNetworkHandles

Arguments

:: [String]
A list of interface names, the display order will be the same as the order in this list

-> IO NetworkHandles

Create a NetworkHandles for the list of networks