Monky.Network.Static

Description

This module should be used for static network configurations

Documentation

getReadWrite :: NetworkHandle -> IO (Maybe (Int, Int))

Get the (read, write) rate of the network interface. This is averaged over the time between calls

data NetworkHandle

Internal handle represanting exactly one interface

getNetworkHandle

Arguments

:: String
Name of the interface to monitor

-> IO NetworkHandle

Get a NetworkHandle

closeNetworkHandle :: NetworkHandle -> IO ()

Close a network handle after it is no longer needed (the device disappeared)