Monky.Outputs.Dzen2

Description

This module provides the output generation for piping into a dzen2 bar

Documentation

data DzenOutput

The output handle for dzen2 pipe

getDzenOut

Arguments

:: Int
The height of your dzen bar in pixel (required for block-drawing)

-> Text
Path to the directory cointaining your .xbm files.

-> IO DzenOutput

Get an output handle for dzen2 formatting