Monky.Outputs.Fallback

Description

Documentation

data WrapOuts

The datatype for wrapping outher outputs

getFallbackOut

Arguments

:: MonkyOutput a

=> IO a
The output to use for non-terminal mode

-> IO WrapOuts

Wrapper for normal outputs that tries to find the best output

This function will check if stdout is a terminal and switch to AsciiOut or UTf8Out depending on the locale

chooseTerminalOut :: IO WrapOuts

Lowerlevel access to guess which terminal out to use based on system encoding