Monky.Prepend

Description

use with: packPrepend Prep instead of normal pack

Documentation

data PrepHandle

The handle used by this module, contains underlying module and string

packPrepend

Arguments

:: Module a

=> String
The String to prepend

-> Int
The refresh rate for this module

-> IO a
The function to get the module

-> IO Modules
The returned handle

Create a module that should be prepended with some string

This allows you to prepend an instance of a module with a fixed String.

For usage look at pack.

data PostHandle

packAppend

Arguments

:: Module a

=> String
The String to prepend

-> Int
The refresh rate for this module

-> IO a
The function to get the module

-> IO Modules
The returned handle

Create a module that should be appended with some string

This allows you to append an instance of a module with a fixed String.

For usage look at pack.