Monky.SSID

Description

This does sometimes fail because the ioctl fails. When this fails is nondeterministic. This module will soon be replaced by netlink based module. This module treats the SSID as C-string, which is wrong but this will not be fixed!! The netlink implementation will replace this.

Documentation

data SSIDHandle

The handle exported by this module

getSSIDHandle :: String -> IO SSIDHandle

Get the handle for this module

getCurrentSSID :: SSIDHandle -> IO String

Get the ssid of the network the device is conncted to or an empty string if it isn't connected