Monky.Template

Description

This module is intended to be used by Monky modules, not for configuration.

This module provides the importLib templateHaskell function to import a C-library easily.

To use this, set the LANGUAGE TemplateHaskell pragma in your module file and include this module.

Usage: Use 'importLib" as top-level declaration in your file. Like:

importLib LibAlsa "libasound.so" []

This will create a data type for the library, and a function to get a handle for this library (data LibAlsa and getLibAlsa). To call your functions use the record syntax on that handle.

Documentation

importLib

Arguments

:: String
The name of the library data type

-> String
The name of the library

-> [(String, String, String)]
The functions in the library (Name, CName, Declaration)

-> Q [Dec]

Import a library

module Foreign.Ptr

module System.Posix.DynamicLinker