Monky.Utility

Description

This module provides utility functions used in monky modules

Documentation

readValue :: File -> IO Int

Read the first line of the file and convert it into an Int

readValues :: File -> IO [Int]

Read the first line of the file and convert the words in it into Ints

fopen :: String -> IO File

open a file read only

type File = Handle

type alias to distinguish system functions from utility

readLine :: File -> IO String

Read the first line of the file

readContent :: File -> IO [String]

Rewind the file descriptor and read the complete file as lines

convertUnit :: Int -> String -> String -> String -> String -> String

Convert a number into a reasonable scale for SI units

findLine :: String -> [String] -> Maybe String

Find a line in a list of Strings

splitAtEvery :: String -> String -> [String]

Split ys at every occurence of xs