Monky.Wifi

Description

Documentation

getCurrentWifi :: SSIDSocket -> Interface -> IO (Maybe String)

Get only the name of the currently connected wifi

getCurrentWifiStats :: SSIDSocket -> Interface -> IO (Maybe WifiStats)

Get the stats of a currently connected wifi network

getInterface :: SSIDSocket -> String -> IO (Maybe Interface)

Get the interface id by name

gotReadable :: SSIDSocket -> Interface -> IO WifiConn

This should be called when the fd returned by getWifiFd got readable

getSSIDSocket :: IO SSIDSocket

Get a netlink socket bound to nl80211 mlme Group

type Interface = Word32

The interface identifier

type SSIDSocket = NL80211Socket

The socket type for this module

getWifiFd :: SSIDSocket -> Fd

get the raw fd for eventing

data WifiStats

Wifi network connection information

Constructors

WifiStats

Fields

wifiChannel :: Word8

wifiRates :: [Word32]

wifiName :: String

wifiFreq :: Word32

wifiMBM :: Word32

data WifiConn

Enum for connection change

Constructors

WifiNone
Nothing changed, connection unrelated message

WifiDisconnect
The current network was disconnectd

WifiConnect WifiStats
A new connection was established