Monky.CPU

Description

Documentation

data CPUHandle

The handle exported by this module

getCPUHandle :: ScalingType -> IO CPUHandle

Create an CPUhandle

getCPUPercent :: CPUHandle -> IO [Int]

Get the cpu usage in percent for each (virtual) cpu

getCPUTemp :: CPUHandle -> IO Int

get current CPU temperature

getCPUMaxScalingFreq :: CPUHandle -> IO Float

this function returns a frequency according th the ScalingType of the handle.

The returned valued will be the max of all (virtual) proceessors on the system.

data ScalingType

Which values should be returned by getCPUFreq

Constructors

ScalingMax
Use the maximum frequencie allowed

ScalingCur
Use the current frequencie used