Monky.Disk

Description

This module allows for some support for btrfs devices. This may be renamed in the future when a general block-device module appears.

Documentation

data DiskHandle

The handle exported by this module

getDiskReadWrite :: DiskHandle -> IO (Int, Int)

Get the read write rates from the disk (in bytes/s)

getDiskFree :: DiskHandle -> IO Int

Get the space left on the disk

getDiskHandle :: String -> Int -> IO DiskHandle

Get the disk handle