Monky.Modules

Description

This module provides the Module class which is used to define a Monky compatible module.

Documentation

data Modules

A wrapper around module instances so they can be put into a list.

Constructors

forall a . Module a => MW a Int

class Module a where

The type class for modules

Minimal complete definition

getText

Methods

getText

Arguments

:: String
The current user

-> a
The handle to this module

-> IO String
The text segment that should be displayed for this module

getFDs

Arguments

:: a
The handle to this module

-> IO [Fd]
The Fds to listen on for events

getEventText

Arguments

:: Fd
The fd that triggered an event

-> String
The current user

-> a
The handle to this module

-> IO String
The text segment that should be displayed

This function is used instead of getText for event triggerd updates.

The default implementation mappes this to getText

pack

Arguments

:: Module a

=> Int
The refresh rate for this module

-> IO a
The function to get a module (get??Handle)

-> IO Modules
The packed module ready to be given to startLoop

Function to make packaging modules easier