Monky.Utility

Description

This module provides utility functions used in monky modules

Documentation

readValue :: File -> IO Int

Read the first line of the file and convert it into an Int

readValues :: File -> IO [Int]

Read the first line of the file and convert the words in it into Ints

fopen :: String -> IO File

open a file read only

fclose :: File -> IO ()

data File

type alias to distinguish system functions from utility

readLine :: LineReadable a => File -> IO a

Read the first line of the file

readContent :: FileReadable a => File -> IO [a]

Rewind the file descriptor and read the complete file as lines

convertUnit :: Int -> String -> String -> String -> String -> String

Convert a number into a fixed length strings

convertUnitB :: Integral a => a -> String -> String

Convert a number into a reasonable scale for binary units

convertUnitSI :: Integral a => a -> String -> String

Convert a number into a reasonable scale for SI units

findLine :: Eq a => [a] -> [[a]] -> Maybe [a]

Find a line in a list of Strings

splitAtEvery :: String -> String -> [String]

Split ys at every occurence of xs

maybeOpenFile :: Maybe String -> IO (Maybe File)

fmap fopen would give Maybe (IO File), this fixes that

sdivBound :: (Integral a, Bounded a) => a -> a -> a

0 save divide, uses maxbound for default value

sdivUBound :: Integral a => a -> a -> a -> a

0 save divide, uses default value